requirements.txt
grpcio
grpcio-tools
mypy
types-protobuf
mypy-protobuf
grpc-stubs
pydantic

black